Rename PersistentUserSettings.js to localStorageManager.js
